requirements.txt
git+https://github.com/m-bain/whisperx
nltk
moviepy
pydub
torch
torchvision
torchaudio
transformers
pydantic
pytest
mysql-connector-python
pyannote.audio
openai-whisper